gemini-service/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
pydantic-settings
